+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 1.6.4 - In progress
 
 - [Feature] Add transparent flipper mockup
+- [Feature] Add broke rpc ble session button to debug settings
 - [CI] Add apk artifacts to Pull Requests
 - [CI] Fix failed build in merge_group